+++ unsupported/Platform.h
@@ -57,4 +57,8 @@ bool Platform_getNetworkIO(NetworkIOData* data);
 
 void Platform_getBattery(double *percent, ACPresence *isOnAC);
 
+void Platform_getHostname(char* buffer, size_t size);
+
+void Platform_getRelease(char** string);
+
 #endif
